move debugger package, clean up init
